src/requirements.txt
absl-py==1.4.0
aiosqlite==0.17.0
anyio==3.6.2
asyncpg==0.27.0
asyncpgsa==0.27.1
attrs==23.1.0
# autopep8 @ file:///opt/conda/conda-bld/autopep8_1650463822033/work
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
DateTime==5.1
fastapi==0.95.1
flatbuffers==23.3.3
fonttools==4.39.3
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.0
h11==0.14.0
# h5py @ file:///opt/concourse/worker/volumes/live/2bf3eb2a-d972-43d4-774c-fc2e1f5380a3/volume/h5py_1637138912046/work
httplib2==0.22.0
httptools==0.5.0
idna==3.4
imageio==2.27.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
iso8601==1.1.0
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mediapipe==0.9.3.0
# mkl-fft==1.3.1
# mkl-random @ file:///opt/concourse/worker/volumes/live/f196a661-8e33-4aac-63bc-efb2ff50e035/volume/mkl_random_1626186080069/work
# mkl-service==2.4.0
networkx==3.1
# numpy @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_e3f85f54-0572-4c0e-b190-4bc4766fc3fenewxarhq/croots/recipe/numpy_and_numpy_base_1652801682879/work
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.3.0.38
packaging==23.1
partition==0.1.2
Pillow==9.5.0
protobuf==3.20.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
# pycodestyle @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_a7riaf725h/croot/pycodestyle_1674267226642/work
pydantic==1.10.7
PyJWT==2.6.0
pyparsing==3.0.9
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.16.2
scipy==1.10.1
# six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.10
starlette==0.26.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
# toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tortoise-orm==0.19.3
typing==3.7.4.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
Werkzeug==2.3.1
zipp==3.15.0
zope.interface==6.0
